Refactor db_config file from models folder to config folder, Remove sql folder
