+++ back/routes/login.js
@@ -1,6 +1,6 @@
 var express = require('express');
 var router = express.Router();
-let db = require('../models/db_config');
+let db = require('../config/db_config');
 var bcrypt = require('bcrypt');
 const jwt = require('jsonwebtoken');
 let secretObj = require('../config/jwt');
