--- back/sql/member_table_change_password_length.sql
@@ -1,2 +1,0 @@
-ALTER TABLE `yebalja_db`.`member_table` 
-CHANGE COLUMN `member_login_pw` `member_login_pw` VARCHAR(60) NOT NULL ;
