--- back/sql/yebalja_db_gisu_table.sql
@@ -1,56 +1,0 @@
--- MySQL dump 10.13  Distrib 8.0.20, for Win64 (x86_64)
---
--- Host: 127.0.0.1    Database: yebalja_db
--- ------------------------------------------------------
--- Server version	8.0.20
-
-/*!40101 SET @OLD_CHARACTER_SET_CLIENT=@@CHARACTER_SET_CLIENT */;
-/*!40101 SET @OLD_CHARACTER_SET_RESULTS=@@CHARACTER_SET_RESULTS */;
-/*!40101 SET @OLD_COLLATION_CONNECTION=@@COLLATION_CONNECTION */;
-/*!50503 SET NAMES utf8 */;
-/*!40103 SET @OLD_TIME_ZONE=@@TIME_ZONE */;
-/*!40103 SET TIME_ZONE='+00:00' */;
-/*!40014 SET @OLD_UNIQUE_CHECKS=@@UNIQUE_CHECKS, UNIQUE_CHECKS=0 */;
-/*!40014 SET @OLD_FOREIGN_KEY_CHECKS=@@FOREIGN_KEY_CHECKS, FOREIGN_KEY_CHECKS=0 */;
-/*!40101 SET @OLD_SQL_MODE=@@SQL_MODE, SQL_MODE='NO_AUTO_VALUE_ON_ZERO' */;
-/*!40111 SET @OLD_SQL_NOTES=@@SQL_NOTES, SQL_NOTES=0 */;
-
---
--- Table structure for table `gisu_table`
---
-
-DROP TABLE IF EXISTS `gisu_table`;
-/*!40101 SET @saved_cs_client     = @@character_set_client */;
-/*!50503 SET character_set_client = utf8mb4 */;
-CREATE TABLE `gisu_table` (
-  `gisu_id` int NOT NULL AUTO_INCREMENT,
-  `program_id` int NOT NULL,
-  `gisu_title` varchar(45) COLLATE utf8_bin NOT NULL,
-  `gisu_start_date` date NOT NULL,
-  `gisu_end_date` date NOT NULL,
-  PRIMARY KEY (`gisu_id`),
-  KEY `program_id_idx` (`program_id`),
-  CONSTRAINT `program_id` FOREIGN KEY (`program_id`) REFERENCES `program_table` (`program_id`) ON DELETE CASCADE
-) ENGINE=InnoDB AUTO_INCREMENT=4 DEFAULT CHARSET=utf8 COLLATE=utf8_bin COMMENT='프로그램에 따른 기수 정보를 갖는 테이블.';
-/*!40101 SET character_set_client = @saved_cs_client */;
-
---
--- Dumping data for table `gisu_table`
---
-
-LOCK TABLES `gisu_table` WRITE;
-/*!40000 ALTER TABLE `gisu_table` DISABLE KEYS */;
-INSERT INTO `gisu_table` VALUES (1,1,'1기 1차','2020-07-04','2020-07-04'),(2,1,'1기 2차','2020-07-04','2020-07-04'),(3,2,'4기','2020-07-04','2020-07-05');
-/*!40000 ALTER TABLE `gisu_table` ENABLE KEYS */;
-UNLOCK TABLES;
-/*!40103 SET TIME_ZONE=@OLD_TIME_ZONE */;
-
-/*!40101 SET SQL_MODE=@OLD_SQL_MODE */;
-/*!40014 SET FOREIGN_KEY_CHECKS=@OLD_FOREIGN_KEY_CHECKS */;
-/*!40014 SET UNIQUE_CHECKS=@OLD_UNIQUE_CHECKS */;
-/*!40101 SET CHARACTER_SET_CLIENT=@OLD_CHARACTER_SET_CLIENT */;
-/*!40101 SET CHARACTER_SET_RESULTS=@OLD_CHARACTER_SET_RESULTS */;
-/*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
-/*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
-
--- Dump completed on 2020-07-04 11:12:52
