--- back/sql/yebalja_db_member_table.sql
@@ -1,54 +1,0 @@
--- MySQL dump 10.13  Distrib 8.0.20, for Win64 (x86_64)
---
--- Host: 127.0.0.1    Database: yebalja_db
--- ------------------------------------------------------
--- Server version	8.0.20
-
-/*!40101 SET @OLD_CHARACTER_SET_CLIENT=@@CHARACTER_SET_CLIENT */;
-/*!40101 SET @OLD_CHARACTER_SET_RESULTS=@@CHARACTER_SET_RESULTS */;
-/*!40101 SET @OLD_COLLATION_CONNECTION=@@COLLATION_CONNECTION */;
-/*!50503 SET NAMES utf8 */;
-/*!40103 SET @OLD_TIME_ZONE=@@TIME_ZONE */;
-/*!40103 SET TIME_ZONE='+00:00' */;
-/*!40014 SET @OLD_UNIQUE_CHECKS=@@UNIQUE_CHECKS, UNIQUE_CHECKS=0 */;
-/*!40014 SET @OLD_FOREIGN_KEY_CHECKS=@@FOREIGN_KEY_CHECKS, FOREIGN_KEY_CHECKS=0 */;
-/*!40101 SET @OLD_SQL_MODE=@@SQL_MODE, SQL_MODE='NO_AUTO_VALUE_ON_ZERO' */;
-/*!40111 SET @OLD_SQL_NOTES=@@SQL_NOTES, SQL_NOTES=0 */;
-
---
--- Table structure for table `member_table`
---
-
-DROP TABLE IF EXISTS `member_table`;
-/*!40101 SET @saved_cs_client     = @@character_set_client */;
-/*!50503 SET character_set_client = utf8mb4 */;
-CREATE TABLE `member_table` (
-  `member_id` int NOT NULL AUTO_INCREMENT,
-  `member_login_nickname` varchar(20) COLLATE utf8_bin NOT NULL,
-  `member_login_pw` varchar(45) COLLATE utf8_bin NOT NULL,
-  `member_login_email` varchar(45) COLLATE utf8_bin NOT NULL,
-  PRIMARY KEY (`member_id`),
-  UNIQUE KEY `member_login_email_UNIQUE` (`member_login_email`),
-  UNIQUE KEY `member_login_nickname_UNIQUE` (`member_login_nickname`)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_bin COMMENT='로그인에 필요한 멤버정보들을 저장하는 테이블.';
-/*!40101 SET character_set_client = @saved_cs_client */;
-
---
--- Dumping data for table `member_table`
---
-
-LOCK TABLES `member_table` WRITE;
-/*!40000 ALTER TABLE `member_table` DISABLE KEYS */;
-/*!40000 ALTER TABLE `member_table` ENABLE KEYS */;
-UNLOCK TABLES;
-/*!40103 SET TIME_ZONE=@OLD_TIME_ZONE */;
-
-/*!40101 SET SQL_MODE=@OLD_SQL_MODE */;
-/*!40014 SET FOREIGN_KEY_CHECKS=@OLD_FOREIGN_KEY_CHECKS */;
-/*!40014 SET UNIQUE_CHECKS=@OLD_UNIQUE_CHECKS */;
-/*!40101 SET CHARACTER_SET_CLIENT=@OLD_CHARACTER_SET_CLIENT */;
-/*!40101 SET CHARACTER_SET_RESULTS=@OLD_CHARACTER_SET_RESULTS */;
-/*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
-/*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
-
--- Dump completed on 2020-07-08 18:30:38
