--- back/sql/yebalja_db_program_table.sql
@@ -1,55 +1,0 @@
--- MySQL dump 10.13  Distrib 8.0.20, for Win64 (x86_64)
---
--- Host: 127.0.0.1    Database: yebalja_db
--- ------------------------------------------------------
--- Server version	8.0.20
-
-/*!40101 SET @OLD_CHARACTER_SET_CLIENT=@@CHARACTER_SET_CLIENT */;
-/*!40101 SET @OLD_CHARACTER_SET_RESULTS=@@CHARACTER_SET_RESULTS */;
-/*!40101 SET @OLD_COLLATION_CONNECTION=@@COLLATION_CONNECTION */;
-/*!50503 SET NAMES utf8 */;
-/*!40103 SET @OLD_TIME_ZONE=@@TIME_ZONE */;
-/*!40103 SET TIME_ZONE='+00:00' */;
-/*!40014 SET @OLD_UNIQUE_CHECKS=@@UNIQUE_CHECKS, UNIQUE_CHECKS=0 */;
-/*!40014 SET @OLD_FOREIGN_KEY_CHECKS=@@FOREIGN_KEY_CHECKS, FOREIGN_KEY_CHECKS=0 */;
-/*!40101 SET @OLD_SQL_MODE=@@SQL_MODE, SQL_MODE='NO_AUTO_VALUE_ON_ZERO' */;
-/*!40111 SET @OLD_SQL_NOTES=@@SQL_NOTES, SQL_NOTES=0 */;
-
---
--- Table structure for table `program_table`
---
-
-DROP TABLE IF EXISTS `program_table`;
-/*!40101 SET @saved_cs_client     = @@character_set_client */;
-/*!50503 SET character_set_client = utf8mb4 */;
-CREATE TABLE `program_table` (
-  `program_id` int NOT NULL AUTO_INCREMENT,
-  `program_title` varchar(45) COLLATE utf8_bin NOT NULL,
-  `program_benefit` varchar(200) COLLATE utf8_bin NOT NULL,
-  `program_content` varchar(500) COLLATE utf8_bin NOT NULL,
-  `program_summary_content` varchar(200) COLLATE utf8_bin NOT NULL,
-  `program_language` varchar(45) COLLATE utf8_bin NOT NULL,
-  PRIMARY KEY (`program_id`)
-) ENGINE=InnoDB AUTO_INCREMENT=3 DEFAULT CHARSET=utf8 COLLATE=utf8_bin COMMENT='	';
-/*!40101 SET character_set_client = @saved_cs_client */;
-
---
--- Dumping data for table `program_table`
---
-
-LOCK TABLES `program_table` WRITE;
-/*!40000 ALTER TABLE `program_table` DISABLE KEYS */;
-INSERT INTO `program_table` VALUES (1,'42 서울','세전 월 100만원 지급','동료평가를 기반으로 교재없이 학습을 진행하는 프로그램','요약메세지','C언어, 서클 진행 후 다른 언어도 가능'),(2,'SSAFY','월 100만원 지급','알고리즘 중심 교육으로 .. 하는 프로그램','요약메세지2','자바, 파이썬');
-/*!40000 ALTER TABLE `program_table` ENABLE KEYS */;
-UNLOCK TABLES;
-/*!40103 SET TIME_ZONE=@OLD_TIME_ZONE */;
-
-/*!40101 SET SQL_MODE=@OLD_SQL_MODE */;
-/*!40014 SET FOREIGN_KEY_CHECKS=@OLD_FOREIGN_KEY_CHECKS */;
-/*!40014 SET UNIQUE_CHECKS=@OLD_UNIQUE_CHECKS */;
-/*!40101 SET CHARACTER_SET_CLIENT=@OLD_CHARACTER_SET_CLIENT */;
-/*!40101 SET CHARACTER_SET_RESULTS=@OLD_CHARACTER_SET_RESULTS */;
-/*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
-/*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
-
--- Dump completed on 2020-07-04 11:12:52
